server/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Telegram bot
python-telegram-bot==20.6

# Security and encryption
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1

# Async support
aiofiles==23.2.1
asyncio==3.4.3

# WebSocket
websockets==12.0

# File handling
python-magic==0.4.27

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging
python-json-logger==2.0.7

# Optional: Rate limiting
slowapi==0.1.9

# Optional: Monitoring
prometheus-client==0.19.0